[CSPM] Avoid recomputing rule evaluation when inputs are unchanged
This PR introduces some hashing of our resolver outputs to being
able to avoid doing unnecessary Rego evaluation if the inputs is
unchanged.

The main motivation is to avoid the heavy lifting of compiling
Rego policies for nothing. This step if the one that shows up
the most in our profiles, while are inputs are most of the time
constants.